Refactor mapping selection steps: - Create common Mapping selection to avoid duplicate code - Share modelMappingsValidation between SyncAction and Mapping Template - Add metadataType and metadataId to ModelMapping and Mapping selection as columns
