+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2021-12-10T07:02:34.252Z\n"
-"PO-Revision-Date: 2021-12-10T07:02:34.252Z\n"
+"POT-Creation-Date: 2021-12-10T09:03:07.587Z\n"
+"PO-Revision-Date: 2021-12-10T09:03:07.588Z\n"
 
 msgid "Data values - Create/update"
 msgstr ""
@@ -35,7 +35,7 @@ msgstr ""
 msgid "Field {{field}} is not valid"
 msgstr ""
 
-msgid "Only can exists a mapping model by dhis2 model"
+msgid "Only can exists a mapping model by metadata and dhis2 model"
 msgstr ""
 
 msgid "All time"
@@ -90,6 +90,11 @@ msgid "Last year"
 msgstr ""
 
 msgid "Last 5 years"
+msgstr ""
+
+msgid ""
+"An error has occurred saving the action.\n"
+"{{error}}"
 msgstr ""
 
 msgid "You need to select one dataValues mapping"
@@ -176,33 +181,6 @@ msgstr ""
 msgid "Connection"
 msgstr ""
 
-msgid "Dhis2 Model"
-msgstr ""
-
-msgid "xMART Table"
-msgstr ""
-
-msgid "Details"
-msgstr ""
-
-msgid "Delete"
-msgstr ""
-
-msgid "Delete Actions?"
-msgstr ""
-
-msgid "Are you sure you want to delete {{total}} model mappings?"
-msgstr ""
-
-msgid "Ok"
-msgstr ""
-
-msgid "Model Mapping"
-msgstr ""
-
-msgid "Import Template"
-msgstr ""
-
 msgid "Selected {{difference}} elements"
 msgstr ""
 
@@ -277,6 +255,9 @@ msgstr ""
 msgid "Import Results"
 msgstr ""
 
+msgid "Ok"
+msgstr ""
+
 msgid "Import"
 msgstr ""
 
@@ -287,6 +268,33 @@ msgid "Messages"
 msgstr ""
 
 msgid "JSON Response"
+msgstr ""
+
+msgid "Dhis2 Model"
+msgstr ""
+
+msgid "Metadata Type"
+msgstr ""
+
+msgid "Metadata Id"
+msgstr ""
+
+msgid "xMART Table"
+msgstr ""
+
+msgid "Details"
+msgstr ""
+
+msgid "Delete"
+msgstr ""
+
+msgid "Delete Actions?"
+msgstr ""
+
+msgid "Are you sure you want to delete {{total}} model mappings?"
+msgstr ""
+
+msgid "Import Template"
 msgstr ""
 
 msgid "Select mapping template"
@@ -331,6 +339,9 @@ msgid "TEI Attributes"
 msgstr ""
 
 msgid "Enrollments"
+msgstr ""
+
+msgid "Model Mapping"
 msgstr ""
 
 msgid "Existed Table"
